debug

+++ bridge_dataset/bridge_dataset_dataset_builder.py
@@ -147,7 +147,7 @@ def _generate_examples(paths) -> Iterator[Tuple[str, Any]]:
                         observation[new_key] = np.zeros_like(example['observations'][i]['images0'])
                 # observation['visual_trajectory'] = list_traj_img[i]
                 # observation['depth'] = depth_image[i]
-                observation['depth'] = np.random.random((256, 256))
+                observation['depth'] = np.random.random((256, 256)).astype(np.float32)
                 observation['visual_trajectory'] = np.zeros_like((256, 256, 3))
                 observation['tcp_point_2d'] = np.array([1.1, 2.1], dtype=np.int32)
                 observation['tcp_point_23'] = np.array([1.1, 2.1, 3.1], dtype=np.int32)
